backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
asyncpg==0.29.0
python-dotenv==1.0.1
pydantic==2.8.2
pydantic-settings==2.4.0
httpx==0.27.2
python-jose[cryptography]==3.3.0
google-generativeai==0.8.3
xgboost==2.1.1
scikit-learn==1.5.2
joblib==1.4.2
numpy==1.26.4
pandas==2.2.2
